Added DatabaseException handling for integrity violations on delete

+++ src/main/java/com/icarogabriel/workshop/services/UserService.java
@@ -2,8 +2,11 @@ package com.icarogabriel.workshop.services;
 
 import com.icarogabriel.workshop.entities.User;
 import com.icarogabriel.workshop.repositories.UserRepository;
+import com.icarogabriel.workshop.services.exceptions.DatabaseException;
 import com.icarogabriel.workshop.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,9 +32,17 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
-    }
 
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+    }
     public User update(Long id, User obj) {
         User entity = userRepository.getReferenceById(id);
         updateData(entity, obj);
